feat: problem scraper service test 수정

+++ src/main/java/com/aloc/aloc/scraper/ProblemScraperService.java
@@ -148,7 +148,7 @@ public class ProblemScraperService {
 	}
 
 	private boolean isNewProblem(String problemNumber, ProblemType problemType) {
-		int problemId = Integer.parseInt(problemNumber);
+		Long problemId = Long.parseLong(problemNumber);
 		return !problemRepository.existsByAlgorithmIdAndProblemType_Course(problemId,
 			problemType.getCourse());
 	}
